+++ src/pages/LoginPage/SignUpPage/SignUpPageStages/AuthorizationPage.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react';
 
 import { useSendUser, useCertifyUser } from '@/apis';
 
-import { Input, Submit } from '@/shared/component';
+import { Input, Button } from '@/shared/component';
 
 import styles from './AuthorizationPage.module.css';
 
@@ -36,7 +36,7 @@ export default function AuthorizationPage({ email, setStage }) {
         </div>
       </div>
       <div className={styles.submit}>
-        <Submit
+        <Button
           btnName='다음으로'
           className={'right'}
           onSubmit={(e) => {
