+++ src/pages/LoginPage/SignUpPage/SignUpPageStages/SignUpFailure/SignUpFailurePage.jsx
@@ -1,8 +1,7 @@
 import { useEffect } from 'react';
 import { Link, useLocation, useNavigate } from 'react-router-dom';
 
-import { BackAppBar, Submit } from '@/shared/component';
-import { Icon } from '@/shared/component';
+import { BackAppBar, Button, Icon } from '@/shared/component';
 
 import styles from './SignUpFailurePage.module.css';
 
@@ -30,7 +29,7 @@ export default function SignUpFailurePage() {
       <div className={styles.pageBottomFrame}>
         <Link to='/signup'>
           <div className={styles.loginButton}>
-            <Submit btnName='뒤로가기' className='right' />
+            <Button btnName='뒤로가기' className='right' />
           </div>
         </Link>
       </div>
