+++ src/pages/SnoroseVerifyPage/SnoroseVerifyPage.jsx
@@ -2,8 +2,7 @@ import { useState } from 'react';
 
 import { TermsPage, VerifyPage, CompletePage } from '@/pages/SnoroseVerifyPage';
 
-import { BackAppBar } from '@/shared/component';
-import { Icon } from '@/shared/component';
+import { BackAppBar, Icon } from '@/shared/component';
 
 import { TITLE_DES } from '@/constants';
