+++ src/shared/component/PTR/PTR.jsx
@@ -1,5 +1,7 @@
-import React, { useEffect, useRef, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
+
 import { Icon } from '@/shared/component';
+
 import styles from './PTR.module.css';
 
 export default function PTR({ children, onRefresh }) {
