+++ src/shared/component/index.js
@@ -1,7 +1,7 @@
 // button
 export { default as ActionButton } from './button/ActionButton/ActionButton';
+export { default as Button } from './button/Button/Button';
 export { default as ResetButton } from './button/ResetButton/ResetButton';
-export { default as Submit } from './button/Submit/Submit';
 export { default as WriteButton } from './button/WriteButton/WriteButton';
 
 // input
