+++ src/shared/component/layout/Footer/Footer.jsx
@@ -1,7 +1,8 @@
+import { Link } from 'react-router-dom';
+
 import { Icon } from '@/shared/component';
 
 import styles from './Footer.module.css';
-import { Link } from 'react-router-dom';
 
 export default function Footer() {
   return (
